[230704] Junit5 설정, 테스트 코드 추가 완료

+++ nw/src/main/java/lab/cherry/nw/controller/UserController.java
@@ -48,6 +48,7 @@ public class UserController {
         return new ResponseEntity<>(userService.getUsers(), new HttpHeaders(), HttpStatus.OK);
     }
 
+    // TODO: 업데이트 필요
     /**
      * [USER] 사용자 업데이트 함수
      *
@@ -63,7 +64,7 @@ public class UserController {
      */
     @PatchMapping("{id}")
     @Operation(summary = "사용자 업데이트", description = "특정 사용자를 업데이트합니다.")
-    public ResponseEntity<?> updateUserById(@PathVariable("id") Long id,
+    public ResponseEntity<?> updateUserById(@PathVariable("id") Integer id,
             @RequestBody UserEntity userDetail) {
 //        Map<String, Object> map = new LinkedHashMap<>();
 
@@ -99,7 +100,7 @@ public class UserController {
      */
     @GetMapping("{id}")
     @Operation(summary = "ID로 사용자 찾기", description = "사용자를 조회합니다.")
-    public ResponseEntity<?> findByUserId(@PathVariable("id") Long id) {
+    public ResponseEntity<?> findByUserId(@PathVariable("id") Integer id) {
         log.info("[UserController] findByUserId...!");
 
 //        final ResultResponse response = ResultResponse.of(SuccessCode.OK, userService.findById(id));
@@ -119,7 +120,7 @@ public class UserController {
      */
     @DeleteMapping("{id}")
     @Operation(summary = "사용자 삭제", description = "사용자를 삭제합니다.")
-    public ResponseEntity<?> deleteUser(@PathVariable("id") Long id) {
+    public ResponseEntity<?> deleteUser(@PathVariable("id") Integer id) {
         log.info("[UserController] deleteUser...!");
         userService.deleteUser(id);
